Small cleanup to geometry functions
