+++ include/coordinates.h
@@ -3,7 +3,7 @@
 #define _COORDINATES_H
 
 #include <iostream>
-#include "geomtypes.h"
+#include "geom.h"
 #include <utility>
 #include <unordered_set>
 
@@ -86,85 +86,8 @@ double degp2meter(double degp, double latp);
 
 double meter2degp(double meter, double latp);
 
-template<typename T>
-void insertIntermediateTiles(const T &points, uint baseZoom, std::unordered_set<TileCoordinates> &tileSet) {
-	Point p2(0, 0);
-	for (auto it = points.begin(); it != points.end(); ++it) {
-		// Line is from p1 to p2
-		Point p1 = p2;
-		p2 = *it;
-
-		// Calculate p2 tile, and mark it
-		double tileXf2 = lon2tilexf(p2.x(), baseZoom), tileYf2 = latp2tileyf(p2.y(), baseZoom);
-		TileCoordinate tileX2 = static_cast<TileCoordinate>(tileXf2), tileY2 = static_cast<TileCoordinate>(tileYf2);
-		tileSet.insert(TileCoordinates(tileX2, tileY2));
-		if (it == points.begin()) continue;	// first point, so no line
-
-		// Calculate p1 tile
-		double tileXf1 = lon2tilexf(p1.x(), baseZoom), tileYf1 = latp2tileyf(p1.y(), baseZoom);
-		TileCoordinate tileX1 = static_cast<TileCoordinate>(tileXf1), tileY1 = static_cast<TileCoordinate>(tileYf1);
-		tileSet.insert(TileCoordinates(tileX1,tileY1));
-
-		// Supercover line algorithm from http://eugen.dedu.free.fr/projects/bresenham/
-		int i;                       // loop counter
-		int ystep, xstep;            // the step on y and x axis
-		int error;                   // the error accumulated during the increment
-		int errorprev;               // *vision the previous value of the error variable
-		int y = tileY1, x = tileX1;  // the line points
-		int ddy, ddx;                // compulsory variables: the double values of dy and dx
-		int dx = tileX2 - tileX1;
-		int dy = tileY2 - tileY1;
-
-		if (dy < 0) { ystep = -1; dy = -dy; } else { ystep = 1; }
-		if (dx < 0) { xstep = -1; dx = -dx; } else { xstep = 1; }
-
-		ddy = 2 * dy;  // work with double values for full precision
-		ddx = 2 * dx;
-		if (ddx >= ddy) {  // first octant (0 <= slope <= 1)
-			// compulsory initialization (even for errorprev, needed when dx==dy)
-			errorprev = error = dx;  // start in the middle of the square
-			for (i=0 ; i < dx ; i++) {  // do not use the first point (already done)
-				x += xstep;
-				error += ddy;
-				if (error > ddx){  // increment y if AFTER the middle ( > )
-					y += ystep;
-					error -= ddx;
-					// three cases (octant == right->right-top for directions below):
-					if (error + errorprev < ddx)  // bottom square also
-						tileSet.insert(TileCoordinates(x, y-ystep));
-					else if (error + errorprev > ddx)  // left square also
-						tileSet.insert(TileCoordinates(x-xstep, y));
-					else {  // corner: bottom and left squares also
-						tileSet.insert(TileCoordinates(x, y-ystep));
-						tileSet.insert(TileCoordinates(x-xstep, y));
-					}
-				}
-				tileSet.insert(TileCoordinates(x, y));
-				errorprev = error;
-			}
-		} else {  // the same as above
-			errorprev = error = dy;
-			for (i=0 ; i < dy ; i++){
-				y += ystep;
-				error += ddx;
-				if (error > ddy){
-					x += xstep;
-					error -= ddy;
-					if (error + errorprev < ddy)
-						tileSet.insert(TileCoordinates(x-xstep, y));
-					else if (error + errorprev > ddy)
-						tileSet.insert(TileCoordinates(x, y-ystep));
-					else{
-						tileSet.insert(TileCoordinates(x-xstep, y));
-						tileSet.insert(TileCoordinates(x, y-ystep));
-					}
-				}
-				tileSet.insert(TileCoordinates(x, y));
-				errorprev = error;
-			}
-		}
-	}
-}
+void insertIntermediateTiles(Linestring const &points, uint baseZoom, std::unordered_set<TileCoordinates> &tileSet);
+void insertIntermediateTiles(Ring const &points, uint baseZoom, std::unordered_set<TileCoordinates> &tileSet);
 
 // the range between smallest y and largest y is filled, for each x
 void fillCoveredTiles(std::unordered_set<TileCoordinates> &tileSet);
@@ -189,6 +112,10 @@ public:
 	Box getTileBox() const;
 	Box getExtendBox() const;
 };
+
+// Round coordinates to integer coordinates of bbox
+// TODO: This should be self-intersection aware!!
+MultiPolygon round_coordinates(TileBbox const &bbox, MultiPolygon const &mp);
 
 #endif //_COORDINATES_H
