+++ include/read_shp.h
@@ -6,7 +6,7 @@
 #include <string>
 #include <vector>
 #include <map>
-#include "geomtypes.h"
+#include "geom.h"
 #include "output_object.h"
 #include "osm_lua_processing.h"
 #include "kaguya.hpp"
